server/requirements.txt
langchain==0.3.0
feedparser==6.0.11
newspaper3k==0.2.8
listparser==0.20
lxml[html_clean]==5.3.0
groq==0.11.0
python-dotenv==1.0.1
scrapegraphai==1.21.0
fastapi[standard]>=0.113.0,<0.114.0
pydantic>=2.7.0,<3.0.0
transformers==4.44.2
langchain_community==0.3.0
beautifulsoup4==4.12.3
httpx==0.27.0
passlib==1.7.4
pymongo==4.8.0
motor==3.5.1
python-jose==3.3.0
python-multipart==0.0.9
apscheduler==3.10.4
aiosmtplib==3.0.2
email-validator==2.2.0
jinja2==3.1.4
langchain-ollama==0.2.0
ollama==0.3.3